Added Strimlit script and modified others
