+++ app1.py
@@ -1,0 +1,95 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Nov 17 01:49:55 2020
+
+@author: tahmeed
+"""
+
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Nov 16 17:48:17 2020
+
+@author: tahmeed
+"""
+
+
+#from flask import Flask, request
+import pandas as pd
+import numpy as np
+import pickle
+#import flasgger
+#from flasgger import Swagger
+import streamlit as st
+
+from PIL import Image
+
+
+#app=Flask(__name__)
+#Swagger(app)
+
+
+pickle_in = open('classifier.pkl', 'rb')
+classifier = pickle.load(pickle_in)
+
+
+#@app.route('/')
+def welcome():
+    return "Welcome ALL"
+
+#@app.route('/predict')
+def predict_note_authentication(variance,skewness,curtosis,entropy):
+    
+    """Let's Authenticate the Banks Note 
+    This is using docstrings for specifications.
+    ---
+    parameters: 
+      - name: variance
+        in: query
+        type: number
+        required: true
+      - name: skewness
+        in: query
+        type: number
+        required: true
+      - name: curtosis
+        in: query
+        type: number
+        required: true
+      - name: entropy
+        in: query
+        type: number
+        required: true
+    responses:
+        200:
+            description: The output values
+        
+    """
+
+
+    prediction = classifier.predict([[variance,skewness,curtosis,entropy]])
+    print(prediction)
+    return prediction
+
+def main():
+    st.title("Bank Authentication")
+    html_temp = """
+    <div style="background-color:tomato;padding:10px">
+    <h2 style="color:white;text-align:center;">Streamlit Bank Authenticator ML App </h2>
+    </div>
+    """
+    st.markdown(html_temp,unsafe_allow_html=True)
+    variance = st.text_input("Variance","Type Here")
+    skewness = st.text_input("skewness","Type Here")
+    curtosis = st.text_input("curtosis","Type Here")
+    entropy = st.text_input("entropy","Type Here")
+    result=""
+    if st.button("Predict"):
+        result=predict_note_authentication(variance,skewness,curtosis,entropy)
+    st.success('The output is {}'.format(result))
+    if st.button("About"):
+        st.text("Lets LEarn")
+        st.text("Built with Streamlit")
+
+
+if __name__=='__main__':
+    main()
