+++ flask_api.py
@@ -1,17 +1,23 @@
 # -*- coding: utf-8 -*-
 """
-Created on Mon Nov 16 03:45:30 2020
+Created on Mon Nov 16 17:48:17 2020
 
 @author: tahmeed
 """
+
 
 from flask import Flask, request
 import pandas as pd
 import numpy as np
 import pickle
+import flasgger
+from flasgger import Swagger
 
 
 app=Flask(__name__)
+Swagger(app)
+
+
 pickle_in = open('classifier.pkl', 'rb')
 classifier = pickle.load(pickle_in)
 
@@ -22,6 +28,32 @@ def welcome():
 
 @app.route('/predict')
 def predict_note_authentication():
+    
+    """Let's Authenticate the Banks Note 
+    This is using docstrings for specifications.
+    ---
+    parameters: 
+      - name: variance
+        in: query
+        type: number
+        required: true
+      - name: skewness
+        in: query
+        type: number
+        required: true
+      - name: curtosis
+        in: query
+        type: number
+        required: true
+      - name: entropy
+        in: query
+        type: number
+        required: true
+    responses:
+        200:
+            description: The output values
+        
+    """
     variance = request.args.get('variance')
     skewness = request.args.get('skewness')
     curtosis = request.args.get('curtosis')
@@ -31,6 +63,21 @@ def predict_note_authentication():
 
 @app.route('/predict_file', methods=["POST"])
 def predict_note_file():
+    
+    """Let's Authenticate the Banks Note 
+    This is using docstrings for specifications.
+    ---
+    parameters: 
+      - name: file
+        in: formData
+        type: file
+        required: true
+
+    responses:
+        200:
+            description: The output values
+        
+    """
     df_test = pd.read_csv(request.files.get("file"))
     prediction = classifier.predict(df_test)
     return "The Predicted Values for the CSV is"+str(list(prediction))
